Address PR review: extract hook, use runCommand, translations, theme zIndex, private pkg
- Extract all useEffects and state into `useStarter()` custom hook
- Replace raw `$e.run` with `runCommand` from `@elementor/editor-v1-adapters`
- Wrap all user-facing strings with `__()` from `@wordpress/i18n`
- Replace hardcoded zIndex 10000 with `theme.zIndex.appBar + 1`
- Add `class` to MutationObserver attributeFilter for panel collapse
- Set package to `private: true`, remove `publishConfig`

Made-with: Cursor
