+++ packages/packages/core/editor-starter/src/components/starter-overlay.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { useCallback, useEffect, useRef, useState } from 'react';
 import {
 	Card,
 	CardActionArea,
@@ -12,146 +11,29 @@ import {
 	type Theme,
 	Typography,
 } from '@elementor/ui';
+import { __ } from '@wordpress/i18n';
 
-import type { StarterConfig } from '../types';
-import { deleteStarterConfig, getAssetUrl, getEditingPanelWidth, getStarterConfig, getTopBarHeight } from '../utils';
-
-interface ElementorChannels {
-	panelElements?: {
-		on: ( event: string, callback: () => void ) => void;
-		off: ( event: string, callback: () => void ) => void;
-	};
-}
-
-function getElementorChannels(): ElementorChannels | undefined {
-	return ( window as unknown as { elementor?: { channels?: ElementorChannels } } ).elementor?.channels;
-}
-
-function dismissStarterApi( config: StarterConfig ) {
-	const apiFetch = ( window as unknown as { wp?: { apiFetch?: ( args: object ) => Promise< unknown > } } ).wp
-		?.apiFetch;
-
-	apiFetch?.( {
-		path: config.restPath,
-		method: 'POST',
-		data: { starter_dismissed: true },
-	} );
-}
+import { useStarter } from '../hooks/use-starter';
+import { getAssetUrl } from '../utils';
 
 export default function StarterOverlay() {
-	const [ config, setConfig ] = useState< StarterConfig | null >( null );
-	const [ isDismissing, setIsDismissing ] = useState( false );
-	const [ panelWidth, setPanelWidth ] = useState( 0 );
-	const [ topOffset, setTopOffset ] = useState( 0 );
-	const dismissedRef = useRef( false );
-
-	useEffect( () => {
-		const activate = () => {
-			const cfg = getStarterConfig();
-
-			if ( cfg ) {
-				setConfig( cfg );
-				setPanelWidth( getEditingPanelWidth() );
-				setTopOffset( getTopBarHeight() );
-			}
-		};
-
-		const onCommandAfter = ( e: Event ) => {
-			const detail = ( e as CustomEvent )?.detail;
-
-			if ( detail?.command === 'editor/documents/attach-preview' ) {
-				activate();
-			}
-		};
-
-		window.addEventListener( 'elementor/commands/run/after', onCommandAfter );
-
-		return () => window.removeEventListener( 'elementor/commands/run/after', onCommandAfter );
-	}, [] );
-
-	const dismiss = useCallback( () => {
-		if ( ! config || dismissedRef.current ) {
-			return;
-		}
-
-		dismissedRef.current = true;
-		setIsDismissing( true );
-
-		dismissStarterApi( config );
-		deleteStarterConfig();
-	}, [ config ] );
-
-	useEffect( () => {
-		if ( ! config ) {
-			return;
-		}
-
-		const channels = getElementorChannels();
-
-		if ( ! channels?.panelElements ) {
-			return;
-		}
-
-		const handleDragStart = () => dismiss();
-
-		channels.panelElements.on( 'element:drag:start', handleDragStart );
-
-		return () => {
-			channels.panelElements?.off( 'element:drag:start', handleDragStart );
-		};
-	}, [ config, dismiss ] );
-
-	useEffect( () => {
-		if ( ! config ) {
-			return;
-		}
-
-		const updateWidth = () => setPanelWidth( getEditingPanelWidth() );
-		const observer = new MutationObserver( updateWidth );
-		const panel = document.querySelector( '.elementor-panel' );
-
-		if ( panel ) {
-			observer.observe( panel, { attributes: true, attributeFilter: [ 'style' ] } );
-		}
-
-		window.addEventListener( 'resize', updateWidth );
-		updateWidth();
-
-		return () => {
-			observer.disconnect();
-			window.removeEventListener( 'resize', updateWidth );
-		};
-	}, [ config ] );
+	const { config, isDismissing, panelWidth, topOffset, dismiss, openAiPlanner, openTemplatesLibrary, onExited } =
+		useStarter();
 
 	if ( ! config ) {
 		return null;
 	}
 
-	const onAiPlannerClick = () => {
-		dismiss();
-
-		if ( config.aiPlannerUrl ) {
-			window.open( config.aiPlannerUrl, '_blank', 'noopener,noreferrer' );
-		}
-	};
-
-	const onTemplatesClick = () => {
-		dismiss();
-
-		const $e = ( window as unknown as { $e?: { run: ( cmd: string, args?: object ) => void } } ).$e;
-		$e?.run( 'library/open' );
-	};
-
 	return (
-		<Slide direction="down" in={ ! isDismissing } mountOnEnter unmountOnExit onExited={ () => setConfig( null ) }>
+		<Slide direction="down" in={ ! isDismissing } mountOnEnter unmountOnExit onExited={ onExited }>
 			<Paper
 				elevation={ 4 }
-				sx={ {
+				sx={ ( theme: Theme ) => ( {
 					position: 'fixed',
 					insetBlockStart: topOffset + 'px',
 					insetInlineStart: panelWidth + 'px',
 					insetInlineEnd: 0,
-					zIndex: 10000,
+					zIndex: theme.zIndex.appBar + 1,
 					display: 'flex',
 					flexDirection: 'column',
 					alignItems: 'center',
@@ -160,11 +42,11 @@ export default function StarterOverlay() {
 					pb: 4,
 					px: 2.5,
 					backgroundColor: '#f9f9fb',
-				} }
+				} ) }
 			>
 				<CloseButton
 					onClick={ dismiss }
-					aria-label="Close"
+					aria-label={ __( 'Close', 'elementor' ) }
 					sx={ ( theme: Theme ) => ( {
 						position: 'absolute',
 						insetBlockStart: theme.spacing( 2 ),
@@ -180,16 +62,16 @@ export default function StarterOverlay() {
 						color: 'text.primary',
 					} }
 				>
-					Start building.
+					{ __( 'Start building.', 'elementor' ) }
 				</Typography>
 
 				<Stack direction="row" spacing={ 3 } justifyContent="center">
 					<Card sx={ { width: 280 } }>
-						<CardActionArea onClick={ onAiPlannerClick }>
+						<CardActionArea onClick={ openAiPlanner }>
 							<CardMedia
 								component="img"
 								image={ getAssetUrl( 'ai-site-planner.png' ) }
-								alt="AI Site Planner"
+								alt={ __( 'AI Site Planner', 'elementor' ) }
 								sx={ {
 									height: 138,
 									p: 1.5,
@@ -200,21 +82,21 @@ export default function StarterOverlay() {
 							/>
 							<CardContent sx={ { textAlign: 'center' } }>
 								<Typography variant="subtitle1" color="text.primary">
-									AI Site Planner
+									{ __( 'AI Site Planner', 'elementor' ) }
 								</Typography>
 								<Typography variant="body2" color="text.secondary" sx={ { mt: 1 } }>
-									Generate your wireframe with AI
+									{ __( 'Generate your wireframe with AI', 'elementor' ) }
 								</Typography>
 							</CardContent>
 						</CardActionArea>
 					</Card>
 
 					<Card sx={ { width: 280 } }>
-						<CardActionArea onClick={ onTemplatesClick }>
+						<CardActionArea onClick={ openTemplatesLibrary }>
 							<CardMedia
 								component="img"
 								image={ getAssetUrl( 'website-templates.png' ) }
-								alt="Website templates"
+								alt={ __( 'Website templates', 'elementor' ) }
 								sx={ {
 									height: 138,
 									p: 1.5,
@@ -225,10 +107,10 @@ export default function StarterOverlay() {
 							/>
 							<CardContent sx={ { textAlign: 'center' } }>
 								<Typography variant="subtitle1" color="text.primary">
-									Website templates
+									{ __( 'Website templates', 'elementor' ) }
 								</Typography>
 								<Typography variant="body2" color="text.secondary" sx={ { mt: 1 } }>
-									Start with a ready-made design
+									{ __( 'Start with a ready-made design', 'elementor' ) }
 								</Typography>
 							</CardContent>
 						</CardActionArea>
